fix(html): also create pagge blocks for plain.html
