+++ test/unit/fixtures/page-block-1-col.plain.html
@@ -1,0 +1,18 @@
+<div><h1 id="single-col-block">Single col block</h1></div>
+<div>
+    <div class="about-bio">
+        <div>
+            <div>
+                <p></p>
+                <img src="https://hlx.blob.core.windows.net/external/c8248664afb141a601d00412936eff2a20f8d5d4#image.png">
+                <p></p>
+                <h2>Jesus Ramirez</h2>
+                <h4>One of the world’s top Photoshop instructors, Jesus has taught millions of creatives like you.</h4>
+                <p></p>
+                <img src="https://hlx.blob.core.windows.net/external/450460dcb5e742c3065ed87d0fd7c117757cb519#image.png">
+                <a href="https://www.behance.net/JRfromPTC">Follow Me</a>
+                <p></p>
+            </div>
+        </div>
+    </div>
+</div>
