+++ test/unit/fixtures/page-block-2-col.plain.html
@@ -1,0 +1,25 @@
+<div><h1 id="double-col-block">Double col block</h1></div>
+<div>
+    <div class="video-text">
+        <div>
+            <div class="video">
+                <a href="https://www.youtube.com/watch?v=kwLTe-ueHA8">https://www.youtube.com/watch?v=kwLTe-ueHA8</a>
+            </div>
+            <div class="text">
+                <h4>Tues, April 21st</h4>
+                <p>Tips for Faster Editing with co-host Ryan Connolly</p>
+                <a href="https://documentcloud.adobe.com/link/track?uri=urn%3Aaaid%3Ascds%3AUS%3A8b451283-daa7-4bbe-8bad-55460af8caf8">(Download PDF)</a>
+            </div>
+        </div>
+        <div>
+            <div class="video">
+                <a href="https://www.youtube.com/watch?v=818roIRrZm0">https://www.youtube.com/watch?v=818roIRrZm0</a>
+            </div>
+            <div class="text">
+                <h4>Tues, April 28th</h4>
+                <p>Vlogging Made Easy with co-host Shameless Maya</p>
+                <a href="https://documentcloud.adobe.com/link/track?uri=urn%3Aaaid%3Ascds%3AUS%3A9f660f99-2e3e-400e-ae71-6d8410b95b8c">(Download PDF)</a>
+            </div>
+        </div>
+    </div>
+</div>
